Extract BaseQueryAssertionsTest
