+++ presto-main/src/test/java/com/facebook/presto/sql/query/TestAggregationOverJoin.java
@@ -16,23 +16,21 @@ package com.facebook.presto.sql.query;
 import org.testng.annotations.Test;
 
 public class TestAggregationOverJoin
+        extends BaseQueryAssertionsTest
 {
     @Test
     public void test()
     {
         // https://github.com/prestodb/presto/issues/10592
-        try (QueryAssertions queryAssertions = new QueryAssertions()) {
-            queryAssertions
-                    .assertQuery(
-                            "WITH " +
-                                    "    t (a, b) AS (VALUES (1, 'a'), (1, 'b')), " +
-                                    "    u (a) AS (VALUES 1) " +
-                                    "SELECT DISTINCT v.a " +
-                                    "FROM ( " +
-                                    "    SELECT DISTINCT a, b " +
-                                    "    FROM t) v " +
-                                    "LEFT JOIN u on v.a = u.a",
-                            "VALUES 1");
-        }
+        assertions().assertQuery(
+                "WITH " +
+                        "    t (a, b) AS (VALUES (1, 'a'), (1, 'b')), " +
+                        "    u (a) AS (VALUES 1) " +
+                        "SELECT DISTINCT v.a " +
+                        "FROM ( " +
+                        "    SELECT DISTINCT a, b " +
+                        "    FROM t) v " +
+                        "LEFT JOIN u on v.a = u.a",
+                "VALUES 1");
     }
 }
