+++ presto-main/src/test/java/com/facebook/presto/sql/query/TestNumericalStability.java
@@ -13,31 +13,15 @@
  */
 package com.facebook.presto.sql.query;
 
-import org.testng.annotations.AfterClass;
-import org.testng.annotations.BeforeClass;
 import org.testng.annotations.Test;
 
 public class TestNumericalStability
+        extends BaseQueryAssertionsTest
 {
-    private QueryAssertions assertions;
-
-    @BeforeClass
-    public void init()
-    {
-        assertions = new QueryAssertions();
-    }
-
-    @AfterClass(alwaysRun = true)
-    public void teardown()
-    {
-        assertions.close();
-        assertions = null;
-    }
-
     @Test
     public void testVariance()
     {
-        assertions.assertQuery(
+        assertions().assertQuery(
                 "SELECT CAST(VAR_SAMP(x + exp(30))/VAR_SAMP(x) AS DECIMAL(3,2)) " +
                         "FROM (VALUES 1.0, 2.0, 3.0, 4.0, 5.0) AS X(x)",
                 "VALUES 1.00");
@@ -46,7 +30,7 @@ public class TestNumericalStability
     @Test
     public void testCovariance()
     {
-        assertions.assertQuery(
+        assertions().assertQuery(
                 "SELECT CAST(COVAR_SAMP(x + exp(30), x + exp(30))/VAR_SAMP(x) AS DECIMAL(3,2)) " +
                         "FROM (VALUES 1.0, 2.0, 3.0, 4.0, 5.0) AS X(x)",
                 "VALUES 1.00");
@@ -55,7 +39,7 @@ public class TestNumericalStability
     @Test
     public void testCorrelation()
     {
-        assertions.assertQuery(
+        assertions().assertQuery(
                 "SELECT CAST(CORR(x + exp(30), x + exp(30)) AS DECIMAL(3,2)) " +
                         "FROM (VALUES 1.0, 2.0, 3.0, 4.0, 5.0) AS X(x)",
                 "VALUES 1.00");
@@ -64,7 +48,7 @@ public class TestNumericalStability
     @Test
     public void testRegressionSlope()
     {
-        assertions.assertQuery(
+        assertions().assertQuery(
                 "SELECT CAST(REGR_SLOPE((x + exp(30)) * 5 + 8, x + exp(30)) AS DECIMAL(3,2)) " +
                         "FROM (VALUES 1.0, 2.0, 3.0, 4.0, 5.0) AS X(x)",
                 "VALUES 5.00");
@@ -73,7 +57,7 @@ public class TestNumericalStability
     @Test
     public void testRegressionIntercept()
     {
-        assertions.assertQuery(
+        assertions().assertQuery(
                 "SELECT CAST(REGR_INTERCEPT((x + exp(20)) * 5 + 8, x + exp(20)) AS DECIMAL(3,2)) " +
                         "FROM (VALUES 1.0, 2.0, 3.0, 4.0, 5.0) AS X(x)",
                 "VALUES 8.00");
