+++ presto-main/src/test/java/com/facebook/presto/sql/query/TestSubqueries.java
@@ -22,8 +22,6 @@ import com.facebook.presto.sql.planner.plan.ValuesNode;
 import com.google.common.collect.ImmutableList;
 import com.google.common.collect.ImmutableMap;
 import org.intellij.lang.annotations.Language;
-import org.testng.annotations.AfterClass;
-import org.testng.annotations.BeforeClass;
 import org.testng.annotations.Test;
 
 import java.util.function.Consumer;
@@ -43,23 +41,9 @@ import static com.facebook.presto.sql.planner.plan.ExchangeNode.Type.REPARTITION
 import static org.testng.Assert.assertEquals;
 
 public class TestSubqueries
+        extends BaseQueryAssertionsTest
 {
     private static final String UNSUPPORTED_CORRELATED_SUBQUERY_ERROR_MSG = "line .*: Given correlated subquery is not supported";
-
-    private QueryAssertions assertions;
-
-    @BeforeClass
-    public void init()
-    {
-        assertions = new QueryAssertions();
-    }
-
-    @AfterClass(alwaysRun = true)
-    public void teardown()
-    {
-        assertions.close();
-        assertions = null;
-    }
 
     @Test
     public void testCorrelatedExistsSubqueriesWithOrPredicateAndNull()
@@ -78,11 +62,11 @@ public class TestSubqueries
     public void testUnsupportedSubqueriesWithCoercions()
     {
         // coercion from subquery symbol type to correlation type
-        assertions.assertFails(
+        assertions().assertFails(
                 "select (select count(*) from (values 1) t(a) where t.a=t2.b limit 1) from (values 1.0) t2(b)",
                 UNSUPPORTED_CORRELATED_SUBQUERY_ERROR_MSG);
         // coercion from t.a (null) to integer
-        assertions.assertFails(
+        assertions().assertFails(
                 "select EXISTS(select 1 from (values (null, null)) t(a, b) where t.a=t2.b GROUP BY t.b) from (values 1, 2) t2(b)",
                 UNSUPPORTED_CORRELATED_SUBQUERY_ERROR_MSG);
     }
@@ -90,17 +74,17 @@ public class TestSubqueries
     @Test
     public void testCorrelatedSubqueriesWithLimit()
     {
-        assertions.assertQuery(
+        assertions().assertQuery(
                 "select (select t.a from (values 1, 2) t(a) where t.a=t2.b limit 1) from (values 1) t2(b)",
                 "VALUES 1");
         // cannot enforce limit 2 on correlated subquery
-        assertions.assertFails(
+        assertions().assertFails(
                 "select (select t.a from (values 1, 2) t(a) where t.a=t2.b limit 2) from (values 1) t2(b)",
                 UNSUPPORTED_CORRELATED_SUBQUERY_ERROR_MSG);
-        assertions.assertQuery(
+        assertions().assertQuery(
                 "select (select sum(t.a) from (values 1, 2) t(a) where t.a=t2.b group by t.a limit 2) from (values 1) t2(b)",
                 "VALUES BIGINT '1'");
-        assertions.assertQuery(
+        assertions().assertQuery(
                 "select (select count(*) from (select t.a from (values 1, 1, null, 3) t(a) limit 1) t where t.a=t2.b) from (values 1, 2) t2(b)",
                 "VALUES BIGINT '1', BIGINT '0'");
         assertExistsRewrittenToAggregationBelowJoin(
@@ -108,7 +92,7 @@ public class TestSubqueries
                 "VALUES true, false",
                 false);
         // TransformCorrelatedScalarAggregationToJoin does not fire since limit is above aggregation node
-        assertions.assertFails(
+        assertions().assertFails(
                 "select (select count(*) from (values 1, 1, 3) t(a) where t.a=t2.b limit 1) from (values 1) t2(b)",
                 UNSUPPORTED_CORRELATED_SUBQUERY_ERROR_MSG);
         assertExistsRewrittenToAggregationBelowJoin(
@@ -122,10 +106,10 @@ public class TestSubqueries
     public void testCorrelatedSubqueriesWithGroupBy()
     {
         // t.a is not a "constant" column, group by does not guarantee single row per correlated subquery
-        assertions.assertFails(
+        assertions().assertFails(
                 "select (select count(*) from (values 1, 2, 3, null) t(a) where t.a<t2.b GROUP BY t.a) from (values 1, 2, 3) t2(b)",
                 "Scalar sub-query has returned multiple rows");
-        assertions.assertQuery(
+        assertions().assertQuery(
                 "select (select count(*) from (values 1, 1, 2, 3, null) t(a) where t.a<t2.b GROUP BY t.a HAVING count(*) > 1) from (values 1, 2) t2(b)",
                 "VALUES null, BIGINT '2'");
         assertExistsRewrittenToAggregationBelowJoin(
@@ -141,7 +125,7 @@ public class TestSubqueries
                 "VALUES false, true",
                 true);
         // t.b is not a "constant" column, cannot be pushed above aggregation
-        assertions.assertFails(
+        assertions().assertFails(
                 "select EXISTS(select 1 from (values (1, 1), (1, 1), (null, null), (3, 3)) t(a, b) where t.a+t.b<t2.b GROUP BY t.a) from (values 1, 2) t2(b)",
                 UNSUPPORTED_CORRELATED_SUBQUERY_ERROR_MSG);
         assertExistsRewrittenToAggregationAboveJoin(
@@ -156,7 +140,7 @@ public class TestSubqueries
                 "select EXISTS(select * from (values 1, 1, 2, 3) t(a) where t.a=t2.b GROUP BY t.a HAVING count(*) > 1) from (values 1, 2) t2(b)",
                 "VALUES true, false",
                 false);
-        assertions.assertQuery(
+        assertions().assertQuery(
                 "select EXISTS(select * from (select t.a from (values (1, 1), (1, 1), (1, 2), (1, 2), (3, 3)) t(a, b) where t.b=t2.b GROUP BY t.a HAVING count(*) > 1) t where t.a=t2.b)" +
                         " from (values 1, 2) t2(b)",
                 "VALUES true, false");
@@ -169,14 +153,14 @@ public class TestSubqueries
     @Test
     public void testCorrelatedLateralWithGroupBy()
     {
-        assertions.assertQuery(
+        assertions().assertQuery(
                 "select * from (values 1, 2) t2(b), LATERAL (select t.a from (values 1, 1, 3) t(a) where t.a=t2.b GROUP BY t.a)",
                 "VALUES (1, 1)");
-        assertions.assertQuery(
+        assertions().assertQuery(
                 "select * from (values 1, 2) t2(b), LATERAL (select count(*) from (values 1, 1, 2, 3) t(a) where t.a=t2.b GROUP BY t.a HAVING count(*) > 1)",
                 "VALUES (1, BIGINT '2')");
         // correlated subqueries with grouping sets are not supported
-        assertions.assertFails(
+        assertions().assertFails(
                 "select * from (values 1, 2) t2(b), LATERAL (select t.a, t.b, count(*) from (values (1, 1), (1, 2), (2, 2), (3, 3)) t(a, b) where t.a=t2.b GROUP BY GROUPING SETS ((t.a, t.b), (t.a)))",
                 UNSUPPORTED_CORRELATED_SUBQUERY_ERROR_MSG);
     }
@@ -184,7 +168,7 @@ public class TestSubqueries
     @Test
     public void testLateralWithUnnest()
     {
-        assertions.assertFails(
+        assertions().assertFails(
                 "SELECT * FROM (VALUES ARRAY[1]) t(x), LATERAL (SELECT * FROM UNNEST(x))",
                 UNSUPPORTED_CORRELATED_SUBQUERY_ERROR_MSG);
     }
@@ -192,7 +176,7 @@ public class TestSubqueries
     @Test
     public void testCorrelatedScalarSubquery()
     {
-        assertions.assertQuery(
+        assertions().assertQuery(
                 "SELECT * FROM (VALUES 1, 2) t2(b) WHERE (SELECT b) = 2",
                 "VALUES 2");
     }
@@ -200,7 +184,7 @@ public class TestSubqueries
     @Test
     public void testCorrelatedSubqueryWithExplicitCoercion()
     {
-        assertions.assertQuery(
+        assertions().assertQuery(
                 "SELECT 1 FROM (VALUES 1, 2) t1(b) WHERE 1 = (SELECT cast(b as decimal(7,2)))",
                 "VALUES 1");
     }
@@ -215,7 +199,7 @@ public class TestSubqueries
                                     anyTree(
                                             node(ValuesNode.class)))));
         }
-        assertions.assertQueryAndPlan(actual, expected,
+        assertions().assertQueryAndPlan(actual, expected,
                 anyTree(
                         node(JoinNode.class,
                                 anyTree(
@@ -233,7 +217,7 @@ public class TestSubqueries
         Consumer<Plan> singleStreamingAggregationValidator = plan -> assertEquals(countSingleStreamingAggregations(plan), 1);
         Consumer<Plan> finalAggregationValidator = plan -> assertEquals(countFinalAggregationNodes(plan), extraAggregation ? 1 : 0);
 
-        assertions.assertQueryAndPlan(actual, expected,
+        assertions().assertQueryAndPlan(actual, expected,
                 anyTree(
                         aggregation(
                                 ImmutableMap.of("COUNT", functionCall("count", ImmutableList.of("NON_NULL"))),
