+++ presto-main/src/test/java/com/facebook/presto/sql/query/TestUnnest.java
@@ -13,37 +13,21 @@
  */
 package com.facebook.presto.sql.query;
 
-import org.testng.annotations.AfterClass;
-import org.testng.annotations.BeforeClass;
 import org.testng.annotations.Test;
 
 public class TestUnnest
+        extends BaseQueryAssertionsTest
 {
-    private QueryAssertions assertions;
-
-    @BeforeClass
-    public void init()
-    {
-        assertions = new QueryAssertions();
-    }
-
-    @AfterClass(alwaysRun = true)
-    public void teardown()
-    {
-        assertions.close();
-        assertions = null;
-    }
-
     @Test
     public void testUnnestArrayRows()
     {
-        assertions.assertQuery(
+        assertions().assertQuery(
                 "SELECT * FROM UNNEST(ARRAY[ROW(1, 1.1), ROW(3, 3.3)], ARRAY[ROW('a', true), ROW('b', false)])",
                 "VALUES (1, 1.1, 'a', true), (3, 3.3, 'b', false)");
-        assertions.assertQuery(
+        assertions().assertQuery(
                 "SELECT x, y FROM (VALUES (ARRAY[ROW(1.0, 2), ROW(3, 4.123)])) AS t(a) CROSS JOIN UNNEST(a) t(x, y)",
                 "VALUES (1.0, 2), (3, 4.123)");
-        assertions.assertQuery(
+        assertions().assertQuery(
                 "SELECT x, y, z FROM (VALUES (ARRAY[ROW(1, 2), ROW(3, 4)])) t(a) CROSS JOIN (VALUES (1), (2)) s(z) CROSS JOIN UNNEST(a) t(x, y)",
                 "VALUES (1, 2, 1), (1, 2, 2), (3, 4, 1), (3, 4, 2)");
     }
@@ -51,23 +35,23 @@ public class TestUnnest
     @Test
     public void testUnnestPreserveColumnName()
     {
-        assertions.assertQuery(
+        assertions().assertQuery(
                 "SELECT x FROM UNNEST(CAST(ARRAY[ROW(1, 'a'), ROW(2, 'b')] as ARRAY(ROW(x int, y varchar))))",
                 "VALUES (1), (2)");
 
-        assertions.assertFails(
+        assertions().assertFails(
                 "SELECT x FROM" +
                         "(VALUES (3)) AS t(x)" +
                         "CROSS JOIN UNNEST(CAST(ARRAY[ROW(1, 'a'), ROW(2, 'b')] as ARRAY(ROW(x int, y varchar))))",
                 ".*Column 'x' is ambiguous.*");
 
-        assertions.assertQuery(
+        assertions().assertQuery(
                 "SELECT t.x FROM" +
                         "(VALUES (3)) AS t(x)" +
                         "CROSS JOIN UNNEST(CAST(ARRAY[ROW(1, 'a'), ROW(2, 'b')] as ARRAY(ROW(x int, y varchar))))",
                 "VALUES (3), (3)");
 
-        assertions.assertQuery(
+        assertions().assertQuery(
                 "SELECT u.x FROM" +
                         "(VALUES (3)) AS t(x)" +
                         "CROSS JOIN UNNEST(CAST(ARRAY[ROW(1, 'a'), ROW(2, 'b')] as ARRAY(ROW(x int, y varchar)))) u",
@@ -77,21 +61,21 @@ public class TestUnnest
     @Test
     public void testUnnestMultiExpr()
     {
-        assertions.assertFails(
+        assertions().assertFails(
                 "SELECT x " +
                         "FROM UNNEST(" +
                         "   CAST(ARRAY[ROW(1, 'a'), ROW(2, 'b')] as ARRAY(ROW(x int, y varchar)))," +
                         "   CAST(ARRAY[ROW(1, 'a'), ROW(2, 'b')] as ARRAY(ROW(x int, y varchar))))",
                 ".*Column 'x' is ambiguous.*");
 
-        assertions.assertQuery(
+        assertions().assertQuery(
                 "SELECT t3 " +
                         "FROM UNNEST(" +
                         "   CAST(ARRAY[ROW(1, 'a'), ROW(2, 'b')] as ARRAY(ROW(x int, y varchar)))," +
                         "   CAST(ARRAY[ROW(3, 'c'), ROW(4, 'd')] as ARRAY(ROW(x int, y varchar)))) t(t1,t2,t3,t4)",
                 "VALUES (3), (4)");
 
-        assertions.assertQuery(
+        assertions().assertQuery(
                 "SELECT x " +
                         "FROM UNNEST(" +
                         "   CAST(ARRAY[ROW(1, 'a'), ROW(2, 'b')] as ARRAY(ROW(a int, b varchar)))," +
